refactoring appLoader
