+++ src/App/App.jsx
@@ -1,9 +1,7 @@
-// import Content from './layouts/content';
 import React from 'react';
 import '../style.css';
 import { Redirect, Route, Switch } from 'react-router-dom';
 import routes from './routes';
-import AppLoader from './components/ui/hoc/appLoader';
 
 const getRoutes = (routes) => {
     return routes.map((route, key) => {
@@ -14,12 +12,10 @@ const getRoutes = (routes) => {
 function App() {
     return (
         <>
-            <AppLoader>
-                <Switch>
-                    {getRoutes(routes)}
-                    <Redirect to="/" />
-                </Switch>
-            </AppLoader>
+            <Switch>
+                {getRoutes(routes)}
+                <Redirect to="/" />
+            </Switch>
         </>
     );
 }
