+++ src/App/store/banner.js
@@ -37,6 +37,7 @@ export const loadBannersList = () => async (dispatch) => {
     }
 };
 export const getBannerByLocation = (location) => (state) => {
+    console.log(state.banner.entities);
     const banner = state.banner.entities.find((banner) => banner.path === location.pathname);
     return banner;
 };
